fix department

+++ aspnet-core/src/GWebsite.AbpZeroTemplate.Core/Models/Product.cs
@@ -11,8 +11,8 @@ namespace GWebsite.AbpZeroTemplate.Core.Models
     public partial class Product : Entity<int>
     {
         public string Name { get; set; }
-        public double Price { get; set; }
-        public double Discount { get; set; }
+        public float Price { get; set; }
+        public float Discount { get; set; }
         public string Summary { get; set; }
         public string AvailableAddress { get; set; }
         public DateTime AddedDate { get; set; }
